spring-review/spring-review/src/main/java/com/example/springreview/service/MemberService.java
package com.example.springreview.service;

import com.example.springreview.domain.Member;
import com.example.springreview.repository.MemoryMemberRepository;

import javax.swing.text.html.Option;
import java.util.List;
import java.util.Optional;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class MemberService {

    private final MemoryMemberRepository memberRepository;

    @Autowired
    public MemberService(MemoryMemberRepository memberRepository) {
        this.memberRepository = memberRepository;
    }

    public Long join(Member member) {
        // 중복확인
        validDuplicateMember(member);
        // 회원가입
        memberRepository.save(member);

        return member.getId();
    }

    private void validDuplicateMember(Member member) {
        /* !중복 확인 */
        memberRepository.findByName(member.getName()).ifPresent(m->{
            throw new IllegalStateException("이미 존재하는 회원입니다.");
        });
    }

    public List<Member> findAll() {
        return memberRepository.findAll();
    }

    public Optional<Member> findOne(Long id) {
        return memberRepository.findById(id);
    }

}
